support optional canonical host and repo to resolve recipe ambiguity
If a project happens to be hosted on a remote recipe repository such
as ELPA, then Straight would complain when verifying the installation
that the remote recipe differs from the local one that points directly
to a local filesystem path. So, to resolve this ambiguity, we
optionally support indicating the canonical host and repo, so that
Straight finds them to be consistent with what it finds on the remote
recipe repositories.