Use a local recipe repository to simplify CI workflows
A recipe repository is a canonical way to describe where to find
package sources and how to build them. We now rely on clients to
simply define recipes for packages of interest (including both project
packages as well as any custom dependencies), and then generate a
local recipe repository for the lifetime of CI. This simplifies the
interfaces and the implementations of the various CI modules.