configure package.el specifically for package-lint
Although this "installability" check of dependencies is already
covered in other CI steps and is superfluous in the linting step,
commonly, client repos with a basic configuration would use upstream
package-lint which does perform this check. In order for this check to
produce accurate results, it makes sense to configure
package.el (which is used internally by package-lint) by adding MELPA
to its list of known repositories so that it can find dependencies
there. This use of package.el is contained to the linter.