show built files in the install step
This allows the user to verify that their recipe is doing what they
expect. For example, the package could still pass all CI checks and
unwittingly have left out the Texinfo manual and any assets (such as
images), or have left out some data that's intended to be bundled with
the package whose use may not be covered by tests.