remove call to `package-initialize` in lint module
This was causing some spooky action at a distance as package.el was
finding and loading packages in the user emacs directory! We do not
want package.el to be doing any package loading, as its purpose is
purely to support (stock) package-lint's installability check, which
relies on its archives being populated so that dependencies can be
found there. The actual loading of installed packages is already
handled by Straight.el, and we'd like it to be *exclusively* handled
by Straight.el to keep everything self-contained, with all effects
manifest within the .elci path.