Use elacarte to manage recipe repository during CI
Formerly, we were defining a "XELPA" local recipe repository for use
during the lifecycle of CI. Elacarte does this for us, so we no longer
need to do it here in an ad hoc way. We now use Elacarte to support
various levels of recipe overrides. Project CI recipes override
project recipes override out-of-the-box Elci recipes.