use new low-level elacarte utility for adding recipes from files
The high level utilities provided by Elacarte now distinguish
"primary" and "pointer" recipes, adding the former and traversing the
latter, and also handle various usage scenarios such as prompting
before replacing existing recipes, and so on. We do not need any of
those features, so using this low-level utility is more appropriate.